refactor: extract race orchestration logic into RaceSetting service
