+++ src/main/kotlin/racingcar/Application.kt
@@ -1,25 +1,11 @@
 package racingcar
 
+import racingcar.service.RaceSetting
+
 fun main() {
     try {
-        runMain()
+        RaceSetting.run()
     } catch (e: IllegalArgumentException) {
         throw e
     }
-
-}
-
-fun runMain() {
-    val carNames = InputHandler.readCarNames()
-    val numberOfRounds = InputHandler.readNumberOfRounds()
-    val cars = Cars.fromNames(carNames)
-
-    OutputView.printRaceStart()
-    repeat(numberOfRounds) {
-        cars.moveAll()
-        OutputView.printRound(cars)
-    }
-
-    val winners = cars.getWinners()
-    OutputView.printWinners(winners)
 }
