+++ src/main/kotlin/racingcar/view/InputHandler.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.view
 
 import camp.nextstep.edu.missionutils.Console
